feat(vault): Add emergency pause (#35)
This PR adds emergency pause functionality to the vault contract. This
allows:
* Pause granularity at following levels:
  * deposit
  * swap
* Pausing a particular pool with the pause granularity.
* Pausing the vault as a whole with pause granularity. This would pause
all the pools.
* Only vault admin is allowed to pause/unpause.

Pausing deposits and swaps is useful in case something goes wrong with a
pool, eg: a hack, so any further malicious actions can be stopped.

Pausing withdrawals isn't allowed, as that leads to centralization
risks.