chore: Oak + Peckshield Audit fixes (#38)
## Oak audit fixes: 2022-12-23 Audit Report - Dexter Updates v0.1
1. Malicious users can prevent stakers from withdrawing LP tokens and
rewards: We have added a reward schedule proposal flow to fix this. Now,
anyone can first only propose a reward schedule which will first have to
be approved by the admin in order to become an actual reward schedule
for an LP token.
2. Owner address validation is not performed: Done
3. Multihop swap lacking multiswap_request validation: Done
4. Ended reward schedules consume unnecessary computation power: We
accept this risk. As we are only adding reward schedules after a review
by the admin. So, we don't expect the array to grow too large. And even
when it does, the current code already skips ended reward schedules, the
major compute actually happens in the lines after the skip. So, skip
doesn't really add much to the compute.
5. UnclaimedRewards query returns incorrect value: Done
6. Removing and adding LP tokens might run out of gas if too many tokens
are registered: Done
7. Querying an arbitrary number of hops might misinform users: Done
8. Final hop of multi-hop swap is inefficient for token transfer: It can
happen that there is only 1 multi-hop request, the current code also
handles that scenario. If we change it, that won't be possible. So, this
isn't an issue.
9. Unlocking liquidity pool tokens too frequently will prevent users
from claiming them: Done
10. Return detailed error message rather than generic overflow error:
Done
11. No events are emitted for important executions: Tx params can be
used for off-chain indexing purposes.
12. Additional funds sent to the contract are lost: One can always
transfer any token to the contract directly even if we block it in some
particular code. So, we do not pay attention to this.
13. Error message is ambiguous: Done
14. General inefficiencies in codebase: Done

## Peckshield audit fixes: 2022-29-11
* Note(5): We don't plan to launch `stable_pool` in our production, we
will be using `stable_5pool`. So, we don't pay attention to that code.
* Note(5)-2: Done
* Note(6): Done. Returning error from query for that particular case
instead of returning wrong results.
* Note(8): Done