Fix for greatest precision based LP minting in the Stableswap pool (#64)
Currently we mint LP tokens in the stableswap pool based on greatest
precision present in the pool.
This is wrong on multiple levels:
- LP Token is minted always with fixed decimals by the Vault contract
(currently 18 decimals) and assume to be greatest precision of pool
assets in the stable pool code. This leads to inconsistency in display
of Stable pool LP token in our frontend /elsewhere (like mintscan etc.)
- LP Token precision should be kept maximum is all cases anyways to
ensure best mathematical results. The current changes improves trade
prices (although small but it adds up) in multiple scenarios. Checking
the code, increasing precision shouldn't have any effect on the contract
calculation speeds since underlying calculations in Decimal256 library
are always with 18 Decimal precision so it's more of an effect of not
losing any precision that the library provides.


This PR aims to fix the LP Token mint by the StableSwap exactly to 18
decimal places i.e. Decimal::DECIMAL_PLACES i.e. the default decimals
places both Decimal and Decimal256 types operate.

This makes the usage of LP token consistent with the Weighted Pool as
well!